Add defineWebApplication function
