+++ packages/web-pkg/src/composables/appDefaults/types.ts
@@ -2,6 +2,7 @@ import { MaybeRef } from '../../utils'
 import { LocationQuery } from '../router'
 import { SpaceResource } from 'web-client/src/helpers'
 import { RouteParams } from 'vue-router'
+
 export interface FileContext {
   path: MaybeRef<string>
   driveAliasAndItem: MaybeRef<string>
@@ -13,5 +14,3 @@ export interface FileContext {
   routeParams: MaybeRef<RouteParams>
   routeQuery: MaybeRef<LocationQuery>
 }
-
-export type AppConfigObject = Record<string, any>
